bluetooth: mesh: move key primitives from mesh to keys header
Commit moves enumeration and definition related to keys
from mesh.h to keys.h
