+++ subsys/bluetooth/mesh/proxy_cli.c
@@ -15,7 +15,6 @@
 #include <zephyr/bluetooth/gatt.h>
 #include <zephyr/bluetooth/mesh.h>
 
-#include "mesh.h"
 #include "net.h"
 #include "rpl.h"
 #include "transport.h"
