+++ subsys/bluetooth/mesh/proxy_srv.c
@@ -19,7 +19,6 @@
 
 #include "common/bt_str.h"
 
-#include "mesh.h"
 #include "net.h"
 #include "rpl.h"
 #include "transport.h"
